* feat: 장바구니 상품 갯수 헤더에 표시 및 구매 바에 표시 (#229)

* feat: 장바구니 상품 갯수 헤더에 표시 및 구매 바에 표시

* style : 장바구니 안내사항

* feat: oauth 로그인 함수 추가 및 oauth_cookie를 oauth sign in 시 받아오도록 설정 (#231)

* feat: oauth 설정 (#235)

* feat: 주문 페이지 api 연동 및 배송지 변경 기능 추가 (#238)

확인함

* Update package.json (#242)

* feat: 주문 페이지에서 결제하기 버튼 추가 (#241)

* feat: 주문 페이지에서 결제하기 버튼 추가

* refactor: api 수정

* feat: 제품 상세 페이지 옵션 선택 및 장바구니 추가 기능  (#243)

* feat: 로그인 페이지 isloading 처리 및 loader 추가

* feat: product Detail PerchaseBar에서 옵션이 단일이면 자동으로 PurchaseItem에 추가되는 기능 추가
productNameData 를 productDetail 페이지에서 PurchaseItem까지 props로 전달하여 옵션이 없으면 ProductName을 보여주도록 함

* feat: menuCategoryList 아이템의 curspor pointer 추가 및 이미지 추가

* feat: product detail 페이지에서 컬러/사이즈 옵션 존재유무에 따른 예외 처리

* feat: purchaseItem 갯수 조절 기능 추가
onQuantityChange를 purchaseQuantityControl까지 내려주어서 증감 구현

* style: purchaseBar style 조정
총 가격은 sheet에서 불러오는 bottomContent가 보여주도록 하면서도 perchaseBar의 구매하기 버튼이 상단에 보여지도록 설정

* feat: product detail의 perchaseBar cart add 기능 추가

* feat: purchaseBar 장바구니 추가 기능 예외처리 및 alertModal이 뜨는 기능 추가

* style: product detail 상세정보 html style이 width에 맞도록 조정

---------

+++ src/actions/cart-service/index.ts
@@ -1,6 +1,10 @@
 'use server';
 import { options } from '@/app/api/auth/[...nextauth]/options';
-import { CartItemType, CartProductItemType } from '@/types/CartDataType';
+import {
+  AddCartItemDataType,
+  CartItemType,
+  CartProductItemType,
+} from '@/types/CartDataType';
 import { CommonResponseType } from '@/types/ResponseDataTypes';
 import { getServerSession } from 'next-auth';
 
@@ -9,8 +13,7 @@ import { revalidateTag } from 'next/cache';
 export const getCartItemData = async (): Promise<CartItemType[]> => {
   const session = await getServerSession(options);
   const token = (await session?.user.accessToken) || session?.user.refreshToken;
-  console.log('리프레쉬 토큰 ', session?.user.refreshToken);
-  console.log('토큰 ', token);
+  if (!token) return [];
   const res = await fetch(`${process.env.BASE_API_URL}/api/v1/cart/user`, {
     method: 'GET',
     headers: {
@@ -54,8 +57,7 @@ export const getProductItem = async (
 export const cartItemCheck = async (cartUuid: string, checked: boolean) => {
   const session = await getServerSession(options);
   const token = (await session?.user.accessToken) || session?.user.refreshToken;
-  console.log('id', cartUuid);
-  console.log('checked', checked);
+  if (!token) return;
   const res = await fetch(`${process.env.BASE_API_URL}/api/v1/cart/checked`, {
     method: 'PUT',
     headers: {
@@ -80,9 +82,7 @@ export const cartItemCheck = async (cartUuid: string, checked: boolean) => {
 export const checkedAllItem = async (checked: boolean) => {
   const session = await getServerSession(options);
   const token = (await session?.user.accessToken) || session?.user.refreshToken;
-  console.log('리프레쉬 토큰 ', session?.user.refreshToken);
-  console.log('토큰 ', token);
-  console.log('장바구니 uuid', checked);
+  if (token) return;
   const res = await fetch(
     `${process.env.BASE_API_URL}/api/v1/cart/checked/all`,
     {
@@ -109,9 +109,7 @@ export const updateCartItemQuantity = async (
 ) => {
   const session = await getServerSession(options);
   const token = (await session?.user.accessToken) || session?.user.refreshToken;
-  console.log('리프레쉬 토큰 ', session?.user.refreshToken);
-  console.log('토큰 ', token);
-  console.log('장바구니 uuid', cartUuid, '수량', quantity);
+  if (!token) return;
   const res = await fetch(`${process.env.BASE_API_URL}/api/v1/cart/quantity`, {
     method: 'PUT',
     headers: {
@@ -134,9 +132,7 @@ export const updateCartItemQuantity = async (
 export const deleteCartItem = async (cartUuid: string) => {
   const session = await getServerSession(options);
   const token = (await session?.user.accessToken) || session?.user.refreshToken;
-  console.log('리프레쉬 토큰 ', session?.user.refreshToken);
-  console.log('토큰 ', token);
-  console.log('장바구니 uuid', cartUuid);
+  if (!token) return;
   const res = await fetch(`${process.env.BASE_API_URL}/api/v1/cart`, {
     method: 'DELETE',
     headers: {
@@ -157,8 +153,7 @@ export const deleteCartItem = async (cartUuid: string) => {
 export const deleteAllCartItem = async () => {
   const session = await getServerSession(options);
   const token = (await session?.user.accessToken) || session?.user.refreshToken;
-  console.log('리프레쉬 토큰 ', session?.user.refreshToken);
-  console.log('토큰 ', token);
+  if (!token) return;
   const res = await fetch(`${process.env.BASE_API_URL}/api/v1/cart/all`, {
     method: 'DELETE',
     headers: {
@@ -172,3 +167,39 @@ export const deleteAllCartItem = async () => {
   }
   revalidateTag('getCartData');
 };
+
+export async function AddCartItemAction(AddCartItemData: AddCartItemDataType) {
+  const session = await getServerSession(options);
+  if (!session)
+    return {
+      success: false,
+      message: '로그인이 필요한 서비스입니다. 로그인 후 다시 시도해주세요.',
+    };
+
+  const token = await session.user.accessToken;
+
+  const payload: Partial<AddCartItemDataType> = { ...AddCartItemData };
+  console.log(payload);
+  try {
+    const response = await fetch(`${process.env.BASE_API_URL}/api/v1/cart`, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json',
+        'Authorization': `Bearer ${token}`,
+      },
+      body: JSON.stringify(payload),
+    });
+
+    if (!response.ok) {
+      const errorData = await response.json();
+      console.error('add cart failed:', errorData);
+      // throw new Error(errorData.message);
+      return { success: errorData.success, message: errorData.message };
+    }
+
+    const data = await response.json();
+    return { success: data.success, message: data.message };
+  } catch (error) {
+    return { success: false, message: '알 수 없는 오류가 발생했습니다.' };
+  }
+}
